requirements.txt
discord.py>=2.3.0
google-genai>=0.3.0
python-dotenv>=1.0.0
APScheduler>=3.10.0
tabulate>=0.9.0
libsql-experimental>=0.0.55
pytz>=2024.1
